TST common tests
