+++ sklearn/metrics/_plot/tests/test_plot_det_curve.py
@@ -2,14 +2,8 @@ import pytest
 import numpy as np
 from numpy.testing import assert_allclose
 
-from sklearn.base import ClassifierMixin
-from sklearn.compose import make_column_transformer
 from sklearn.datasets import load_iris
-from sklearn.exceptions import NotFittedError
 from sklearn.linear_model import LogisticRegression
-from sklearn.pipeline import make_pipeline
-from sklearn.preprocessing import StandardScaler
-from sklearn.tree import DecisionTreeClassifier
 
 from sklearn.metrics import det_curve
 from sklearn.metrics import plot_det_curve
@@ -24,42 +18,6 @@ def data():
 def data_binary(data):
     X, y = data
     return X[y < 2], y[y < 2]
-
-
-def test_plot_det_curve_error_non_binary(pyplot, data):
-    X, y = data
-    clf = DecisionTreeClassifier()
-    clf.fit(X, y)
-
-    msg = "DecisionTreeClassifier should be a binary classifier"
-    with pytest.raises(ValueError, match=msg):
-        plot_det_curve(clf, X, y)
-
-
-@pytest.mark.parametrize(
-    "response_method, msg",
-    [("predict_proba", "response method predict_proba is not defined in "
-                       "MyClassifier"),
-     ("decision_function", "response method decision_function is not defined "
-                           "in MyClassifier"),
-     ("auto", "response method decision_function or predict_proba is not "
-              "defined in MyClassifier"),
-     ("bad_method", "response_method must be 'predict_proba', "
-                    "'decision_function' or 'auto'")])
-def test_plot_det_curve_error_no_response(
-    pyplot, data_binary, response_method, msg
-):
-    X, y = data_binary
-
-    class MyClassifier(ClassifierMixin):
-        def fit(self, X, y):
-            self.classes_ = [0, 1]
-            return self
-
-    clf = MyClassifier().fit(X, y)
-
-    with pytest.raises(ValueError, match=msg):
-        plot_det_curve(clf, X, y, response_method=response_method)
 
 
 @pytest.mark.parametrize(
@@ -124,35 +82,3 @@ def test_plot_det_curve(
     )
     assert viz.ax_.get_ylabel() == expected_ylabel
     assert viz.ax_.get_xlabel() == expected_xlabel
-
-
-@pytest.mark.parametrize(
-    "clf", [LogisticRegression(),
-            make_pipeline(StandardScaler(), LogisticRegression()),
-            make_pipeline(make_column_transformer((StandardScaler(), [0, 1])),
-                          LogisticRegression())])
-def test_plot_det_curve_not_fitted_errors(pyplot, data_binary, clf):
-    X, y = data_binary
-    with pytest.raises(NotFittedError):
-        plot_det_curve(clf, X, y)
-    clf.fit(X, y)
-    disp = plot_det_curve(clf, X, y)
-    assert clf.__class__.__name__ in disp.line_.get_label()
-    assert disp.estimator_name == clf.__class__.__name__
-
-
-def test_plot_det_curve_estimator_name_multiple_calls(pyplot, data_binary):
-    # non-regression test checking that the `name` used when calling
-    # `plot_det_curve` is used as well when calling `disp.plot()`
-    X, y = data_binary
-    clf_name = "my hand-crafted name"
-    clf = LogisticRegression().fit(X, y)
-    disp = plot_det_curve(clf, X, y, name=clf_name)
-    assert disp.estimator_name == clf_name
-    pyplot.close("all")
-    disp.plot()
-    assert clf_name in disp.line_.get_label()
-    pyplot.close("all")
-    clf_name = "another_name"
-    disp.plot(name=clf_name)
-    assert clf_name in disp.line_.get_label()
